Simplify fog_client files and dir handling
This makes idempotent_directory unnecessary. Its concept is questionable anyways, because even a pure read request that returns a 404 can create a directory on the server. This problem still exists, but it is now made more obvious and easier to fix, because the problem is now visible in the fog_client itself.